#2786 - Plugging in JS editors
- Move more components to auto-configuration
- Make brat annotation serializer accessible by editor plugins
