+++ inception/inception-brat-editor/src/main/java/de/tudarmstadt/ukp/clarin/webanno/brat/annotation/BratAnnotationEditor.java
@@ -82,6 +82,7 @@ import de.tudarmstadt.ukp.clarin.webanno.brat.message.VisualOptions;
 import de.tudarmstadt.ukp.clarin.webanno.brat.metrics.BratMetrics;
 import de.tudarmstadt.ukp.clarin.webanno.brat.metrics.BratMetrics.RenderType;
 import de.tudarmstadt.ukp.clarin.webanno.brat.render.BratSerializer;
+import de.tudarmstadt.ukp.clarin.webanno.brat.render.BratSerializerImpl;
 import de.tudarmstadt.ukp.clarin.webanno.brat.resource.BratCssReference;
 import de.tudarmstadt.ukp.clarin.webanno.brat.resource.BratResourceReference;
 import de.tudarmstadt.ukp.clarin.webanno.support.JSONUtil;
@@ -107,6 +108,7 @@ public class BratAnnotationEditor
     private @SpringBean BratMetrics metrics;
     private @SpringBean BratAnnotationEditorProperties bratProperties;
     private @SpringBean EditorAjaxRequestHandlerExtensionPoint handlers;
+    private @SpringBean BratSerializer bratSerializer;
 
     private WebMarkupContainer vis;
     private AbstractAjaxBehavior requestHandler;
@@ -260,7 +262,7 @@ public class BratAnnotationEditor
     {
         GetCollectionInformationResponse info = new GetCollectionInformationResponse();
         if (getModelObject().getProject() != null) {
-            info.setEntityTypes(BratSerializer.buildEntityTypes(getModelObject().getProject(),
+            info.setEntityTypes(BratSerializerImpl.buildEntityTypes(getModelObject().getProject(),
                     getModelObject().getAnnotationLayers(), annotationService));
             info.getVisualOptions()
                     .setArcBundle(getModelObject().getPreferences().isCollapseArcs()
@@ -426,7 +428,7 @@ public class BratAnnotationEditor
         AnnotatorState aState = getModelObject();
 
         return render(aCas, aState.getWindowBeginOffset(), aState.getWindowEndOffset(),
-                new BratSerializer(bratProperties));
+                bratSerializer);
     }
 
     private String bratInitCommand()
@@ -454,7 +456,7 @@ public class BratAnnotationEditor
         LOG.trace("[{}][{}] bratLoadCollectionCommand", getMarkupId(), vis.getMarkupId());
 
         GetCollectionInformationResponse response = actionGetCollectionInformation();
-        response.setEntityTypes(BratSerializer.buildEntityTypes(getModelObject().getProject(),
+        response.setEntityTypes(BratSerializerImpl.buildEntityTypes(getModelObject().getProject(),
                 getModelObject().getAnnotationLayers(), annotationService));
         String json = toJson(response);
