+++ inception/inception-brat-editor/src/main/java/de/tudarmstadt/ukp/clarin/webanno/brat/render/model/EntityType.java
@@ -17,7 +17,7 @@
  */
 package de.tudarmstadt.ukp.clarin.webanno.brat.render.model;
 
-import static de.tudarmstadt.ukp.clarin.webanno.brat.render.BratSerializer.abbreviate;
+import static de.tudarmstadt.ukp.clarin.webanno.brat.render.BratSerializerImpl.abbreviate;
 import static java.util.Arrays.asList;
 
 import java.util.List;
