+++ inception/inception-brat-editor/src/main/ts/brat/brat.ts
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 import { Ajax } from "./ajax/Ajax";
-import { DiamAjax } from "@inception-project/inception-diam/diam/Diam";
+import { DiamAjax } from "@inception-project/inception-diam";
 import { AnnotatorUI } from "./annotator_ui/AnnotatorUI";
 import { Dispatcher } from "./dispatcher/Dispatcher";
 import { Visualizer } from "./visualizer/Visualizer";
